requirements.txt
beautifulsoup4==4.11.1
colorama==0.4.6
Requests==2.32.4
tldextract==5.1.2
urllib3==1.26.9
html2text==2020.1.16
anthropic>=0.34.0

# Query Optimizer Dependencies
langchain>=0.2.0
langchain-anthropic>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
